fix: debug panel should be removed from dom on dialog close instead of hidden
