+++ packages/react-components/src/components/cesium-map/debug/wfs.tsx
@@ -1,101 +1,64 @@
 import { get } from 'lodash';
-import React, { useEffect, useState, useMemo } from 'react';
+import React, { useMemo } from 'react';
 import { Tooltip } from '@map-colonies/react-core';
 import { Box } from '../../box';
-import { ICesiumWFSLayer } from '../layers/wfs.layer';
-import { useCesiumMap } from '../map';
+import { IActiveFeatureTypes } from './debug-panel';
 
 import './wfs.css';
 
-interface IFeatureTypeMetadata {
-  id: string;
-  items: number;
-  total: number;
-  cache: number;
-  currentZoomLevel: number;
-  featureStructure: Record<string, unknown>;
-}
-
-type IActiveFeatureTypes = IFeatureTypeMetadata & {
-  zoomLevel: number;
-};
-
 interface IWFSProps {
+  featureTypes: IActiveFeatureTypes[];
   locale?: { [key: string]: string };
 }
 
-export const WFS: React.FC<IWFSProps> = ({ locale }) => {
-  const mapViewer = useCesiumMap();
-  const [featureTypes, setFeatureTypes] = useState<IActiveFeatureTypes[]>([]);
+export const WFS: React.FC<IWFSProps> = ({ featureTypes, locale }) => {
   const title = useMemo(() => get(locale, 'WFS_TITLE') ?? 'Data Layers', [locale]);
   const cacheLabel = useMemo(() => get(locale, 'WFS_CACHE') ?? 'Cache', [locale]);
   const extentLabel = useMemo(() => get(locale, 'WFS_EXTENT') ?? 'Extent', [locale]);
+  const noDataLayers = useMemo(() => get(locale, 'NO_DATA_LAYERS') ?? 'No layers found', [locale]);
 
-  useEffect(() => {
-    if (!mapViewer.layersManager) return;
+  const header = useMemo(() => {
+    return (
+      <Box className="title">{title}</Box>
+    );
+  }, [title]);
 
-    const handleDataLayerUpdated = (dataLayers: ICesiumWFSLayer[], LayerId?: string | undefined): void => {
-      dataLayers.forEach((layer: ICesiumWFSLayer): void => {
-        if (LayerId !== undefined && LayerId !== layer.meta.id) {
-          return;
+  const content = useMemo(() => {
+    return (
+      <>
+        {
+          featureTypes.length > 0 ? (
+            featureTypes.map((type, index) => (
+              <Box key={index} className="featureType">
+                <Tooltip content={`${type.featureStructure.aliasLayerName as string} ${type.id} (${String(type.zoomLevel)})`}>
+                  <Box className={`name ${type.currentZoomLevel < type.zoomLevel ? 'warning blinking' : type.total === -1 ? 'error blinking' : ''}`}>
+                    {type.featureStructure.aliasLayerName as string} {type.id} ({String(type.zoomLevel)}):
+                  </Box>
+                </Tooltip>
+                <Box className="info">
+                  <Box>
+                    {cacheLabel}: {type.cache ?? 0}
+                  </Box>
+                  {type.total > 0 && (
+                    <Box className="spacer">
+                      {extentLabel}: {type.items} / {type.total}
+                    </Box>
+                  )}
+                </Box>
+              </Box>
+            ))
+          ) : (
+            <Box>{noDataLayers}</Box>
+          )
         }
-
-        const { options, meta } = layer;
-        const { zoomLevel } = options;
-        const { id, items, total, cache, currentZoomLevel, featureStructure } = meta as unknown as IFeatureTypeMetadata;
-
-        setFeatureTypes((prevFeatureTypes) => {
-          const existingIndex = prevFeatureTypes.findIndex((type) => type.id === id);
-          if (existingIndex >= 0) {
-            if (
-              JSON.stringify(prevFeatureTypes[existingIndex]) !==
-              JSON.stringify({ id, items, total, cache, currentZoomLevel, featureStructure, zoomLevel })
-            ) {
-              const updatedFeatureTypes = [...prevFeatureTypes];
-              updatedFeatureTypes[existingIndex] = { id, items, total, cache, currentZoomLevel, featureStructure, zoomLevel };
-              return updatedFeatureTypes;
-            }
-          } else {
-            return [...prevFeatureTypes, { id, items, total, cache, currentZoomLevel, featureStructure, zoomLevel }];
-          }
-          return prevFeatureTypes;
-        });
-      });
-
-      const activeDataLayerIds = new Set(mapViewer.layersManager?.dataLayerList.map((layer) => layer.meta.id));
-
-      setFeatureTypes((prevFeatureTypes) => prevFeatureTypes.filter((type) => activeDataLayerIds.has(type.id)));
-    };
-
-    mapViewer.layersManager.addDataLayerUpdatedListener(handleDataLayerUpdated);
-
-    return () => {
-      mapViewer.layersManager?.removeDataLayerUpdatedListener(handleDataLayerUpdated);
-    };
-  }, [mapViewer.layersManager?.dataLayerList]);
+      </>
+    );
+  }, [featureTypes]);
 
   return (
     <Box className="wfsContainer">
-      <Box className="title">{title}</Box>
-      {featureTypes.map((type, index) => (
-        <Box key={index} className="featureType">
-          <Tooltip content={`${type.featureStructure.aliasLayerName as string} ${type.id} (${String(type.zoomLevel)})`}>
-            <Box className={`name ${type.currentZoomLevel < type.zoomLevel ? 'warning blinking' : type.total === -1 ? 'error blinking' : ''}`}>
-              {type.featureStructure.aliasLayerName as string} {type.id} ({String(type.zoomLevel)}):
-            </Box>
-          </Tooltip>
-          <Box className="info">
-            <Box>
-              {cacheLabel}: {type.cache ?? 0}
-            </Box>
-            {type.total > 0 && (
-              <Box className="spacer">
-                {extentLabel}: {type.items} / {type.total}
-              </Box>
-            )}
-          </Box>
-        </Box>
-      ))}
+      {header}
+      {content}
     </Box>
   );
 };
